@@ -64,7 +64,7 @@
 		"sprintf-js": "^1.1.2",
 		"telnet-socket": "^0.2.3",
 		"title-case": "^3.0.3",
-		"ws": "^8.2.0",
+		"ws": "^8.2.1",
 		"xterm": "^4.13.0",
 		"xterm-addon-fit": "^0.5.0",
 		"xterm-addon-unicode11": "^0.2.0",
@@ -77,14 +77,14 @@
 		"@types/browserify": "^12.0.37",
 		"@types/express": "^4.17.13",
 		"@types/fs-extra": "^9.0.12",
-		"@types/node": "^16.7.1",
+		"@types/node": "^16.7.8",
 		"@types/nodemailer": "^6.4.4",
 		"@types/nodemailer-smtp-transport": "^2.7.5",
 		"@types/sprintf-js": "^1.1.2",
 		"@types/ws": "^7.4.7",
 		"browserify": "^17.0.0",
 		"eslint": "^7.32.0",
-		"typescript": "^4.3.5",
+		"typescript": "^4.4.2",
 		"watchify": "^4.0.0"
 	},
 	"types": "src",
