@@ -70,9 +70,8 @@ let idle, reconnect, lurking;
 let tbt = 1;
 window.onresize = () => {
     if (!wall.hidden) {
-        if (!wall.style.zoom)
-            setImmediate(() => window.dispatchEvent(new Event('resize')));
-        wall.style.zoom = `${90 / (splash.clientWidth / window.innerWidth)}%`;
+        const zoom = Math.trunc(80 / (splash.clientWidth / window.innerWidth));
+        splash.style.transform = `scale(${zoom}%,${zoom}%) translate(0%,0%)`;
     }
     if (!pid && !wpid)
         return;
@@ -375,10 +374,6 @@ function XT(data) {
         audioTune.load();
         audioTune.play().catch(err => { });
     }
-    function wall(msg) {
-        if (pid)
-            fetch(`${app}/player/${pid}/wall?msg=${msg}`, { method: 'POST' });
-    }
 }
 function lurk() {
     if (terminal.hidden) {
@@ -402,9 +397,11 @@ function lurk() {
                 }
                 watch.selectedIndex = -1;
                 el.blur();
+                setImmediate(() => window.dispatchEvent(new Event('resize')));
                 el.onchange = (ev) => {
                     let watch = ev.target;
                     wpid = parseInt(watch[watch.selectedIndex].value);
+                    wall.hidden = true;
                     terminal.hidden = false;
                     term = new xterm_1.Terminal({
                         bellSound: BELL_SOUND, bellStyle: 'sound', cursorBlink: false, scrollback: 0,
@@ -439,6 +436,7 @@ function lurk() {
                                 term.dispose();
                                 wpid = 0;
                                 terminal.hidden = true;
+                                wall.hidden = false;
                                 lurk();
                             };
                             socket.onerror = (ev) => {
